requirements.txt
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Dash stack
dash==2.17.0
dash-bootstrap-components==1.5.0
plotly==5.19.0

# Data stack
pandas==2.2.1
yfinance==0.2.37